+++ compiler/damlc/daml-compiler/src/DA/Daml/Compiler/Repl.hs
@@ -33,10 +33,10 @@ import qualified DA.Daml.LF.TypeChecker as LF
 import DA.Daml.LF.Ast.Optics (packageRefs)
 import qualified DA.Daml.LF.ReplClient as ReplClient
 import DA.Daml.LFConversion (convertModule)
-import DA.Daml.LFConversion.UtilLF (buildPackage)
 import DA.Daml.Options.Types
 import qualified DA.Daml.Preprocessor.Records as Preprocessor
 import DA.Daml.UtilGHC
+import DA.Daml.UtilLF (buildPackage)
 import Data.Bifunctor (first)
 import Data.Functor.Alt
 import Data.Functor.Bind
