+++ compiler/damlc/daml-ide/BUILD.bazel
@@ -36,7 +36,7 @@ da_haskell_library(
     deps = [
         "//compiler/daml-lf-ast",
         "//compiler/damlc/daml-ide-core",
-        "//compiler/damlc/daml-lf-conversion",
+        "//compiler/damlc/daml-lf-util",
         "//compiler/damlc/daml-rule-types",
         "//compiler/damlc/daml-visual",
         "//libs-haskell/da-hs-base",
