+++ compiler/damlc/daml-lf-conversion/BUILD.bazel
@@ -39,6 +39,7 @@ da_haskell_library(
         "//compiler/daml-lf-proto",
         "//compiler/daml-lf-tools",
         "//compiler/damlc/daml-ghc-util",
+        "//compiler/damlc/daml-lf-util",
         "//libs-haskell/da-hs-base",
     ],
 )
